First commit for an events related feature
The events are loaded after a track is loaded (tied to the SoundContainer as for the lyrics so this could/should be changed at one point).
The events provider is songkick and we are using their API.
A click on an event opens the related songkick's page where we can find more details on the venue, the adress etc.
This is a first step as we could think about
- filtering events for a particular place/city
- being able to access the events of an artist from the artist page
- find events in a geo area
- generate playlist based on upcoming events in an area or a venue